// resources/js/pages/attendance/types/attendance.ts
export interface Attendance {
    id: string;
    timeIn: string;
    timeOut: string;
    breakTime: string;
    attendanceStatus: string;
    attendanceDate: string;
    employee_name: string;
    employeeid: string;
    picture: string;
    department: string;
    position: string;
    session: string;
}